[feat] create ProfileImage upload component
